fix(#279): replace QuickExerciseModal with AddExerciseModal on quick add
Remove the redundant QuickExerciseModal that was acting as a proxy to
AddExerciseModal. The thunderbolt quick-add button now opens AddExerciseModal
directly. On exercise selection a workout is created (or reused) and the
exercise is added immediately - with optional copy-from-last support.

Closes #279
